refactor(voice/factory): split factory.py into subpackage [AP-16, T1.3]
Closes anti-pattern #16 hygiene for factory.py (1744 LOC, mixed
responsibilities — exception types + 7 diagnostic helpers + boot
cascade orchestration + 5 component constructors + bypass-strategy
list + 460-LOC orchestrator function + sys.modules dispatch hack).
Pure refactor: zero behaviour change.

Layout per godfile-splits mission Part 3 / T03:

  voice/factory/
    __init__.py     — orchestrator (create_voice_pipeline, ~460 LOC)
                      + VoiceBundle + logger + re-exports of every
                      public-by-history symbol (17 names)
    _capture.py     — _resolve_platform_key + _build_bypass_strategies
    _playback.py    — _create_vad + _create_piper_tts + _create_kokoro_tts
    _validate.py    — VoiceFactoryError + VoicePermissionError + 5
                      preflight gates (_maybe_check_mic_permission,
                      _maybe_check_llm_reachable, _run_boot_preflight,
                      _run_vchl_boot_cascade, _detect_voice_clarity_active)
                      + _create_stt + _create_wake_word_stub
    _diagnostics.py — 7 read-only platform probes
                      (_maybe_log_pipewire_status, _maybe_log_alsa_ucm_status,
                       _maybe_log_macos_diagnostics,
                       _maybe_log_recent_audio_etw_events,
                       _maybe_start_audio_service_watchdog) +
                      _emit_capture_apo_detection +
                      _emit_linux_capture_apo_detection

Critical surgery (per spec §B3): replaced the 6 ``_self.X(...)`` sites
in create_voice_pipeline (formerly using
``_self = sys.modules[__name__]`` for late-bound dispatch) with explicit
imports from the appropriate sub-modules:

  Line 840  _self._create_vad(vad_path)         → _create_vad(vad_path)
  Line 852  _self._create_stt(language)         → _create_stt(language)
  Line 888  _self._create_piper_tts(models_dir) → _create_piper_tts(models_dir)
  Line 892  _self._create_kokoro_tts(...)       → _create_kokoro_tts(...)
  Line 916  _self._create_wake_word_stub        → _create_wake_word_stub
  Line 1097 _self._create_vad(vad_path)         → _create_vad(vad_path)

Removed the ``_self = sys.modules[__name__]`` line entirely. The
dispatch was a pre-CLAUDE.md-AP-20 trick to make module-level
monkeypatches reflect inside ``create_voice_pipeline``; with explicit
package-level imports + the stable ``__init__.py`` re-export surface,
test patches on ``sovyx.voice.factory._create_*`` resolve via the
package namespace and continue to work.

Test patch migration (anti-pattern #20):

  Wave 3 recon reported 10 test patches (all on `create_voice_pipeline`).
  At HEAD a fresh grep confirmed 15 patches against
  `sovyx.voice.factory.create_voice_pipeline` in `test_voice_routes.py`.
  Since the function lives at the package root post-split, all 15
  patches keep working unchanged — the package-level binding IS the
  function.

  Source-level grep tests in 2 files (`test_factory_ring_init_events.py`,
  `test_v0_22_4_to_v0_23_0_migration.py`) referenced the literal path
  ``src/sovyx/voice/factory.py`` and broke because the file moved to
  ``factory/__init__.py``. Migrated both to read the new path.

Validation:
  * 17-name smoke import: OK
  * ruff check + format clean
  * mypy strict: 416 source files, 0 issues
  * bandit clean
  * pytest full suite: 11768 passed, 25 skipped, 57 deselected
    (identical to pre-split baseline)

Refs: docs-internal/missions/MISSION-voice-godfile-splits-v0.24.1.md
      Part 3 / T03 + Blocker B3 (`_self` surgery)
      docs-internal/missions/MISSION-voice-final-skype-grade-2026.md
      Phase 1 / T1.3
      CLAUDE.md anti-pattern #16, #20
